requirements.txt
requests
pyyaml
feedparser
markdown
jinja2
python-dateutil
